requirements.txt
aiofiles==24.1.0
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==43.0.0
diff-match-patch==20230430
Django==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-import-export==4.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
m3u8==6.0.0
nest-asyncio==1.6.0
oauthlib==3.2.2
orjson==3.10.6
packaging==24.1
praw==7.7.1
prawcore==2.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==1.3.1
rpds-py==0.20.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tablib==3.5.0
tqdm==4.66.5
tweepy==4.14.0
twitter-api-client==0.10.22
tzdata==2024.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.2
uvloop==0.19.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
